Show user photo for diary, notes and user management
The profile photo is now shown for diary and notes application. Click will open the edit account setting page.
The account setting page justs shows the photo without interaction.
Simplified headers for diary and notes.
Fixed notes bug if encryption key is wrong, now all notes are shown but with a ? in the title because they cannot be decrypted.